work in progress

+++ .dotfiles/.zshrc
@@ -1,38 +1,50 @@
+if [ "$(hostname -f | cut -d'.' -f2,3)" = "42.fr" ]; then
+    export AT_42=true
+fi
+
 ######################
 #        USER        #
 ######################
 
-if [ -z "$USER" ]; then
-  export USER="$(id -un)"
-fi
-if [ -z "$FULLNAME" ]; then
-  export FULLNAME="$(id -F)"
-fi
-if [ -z "$MAIL" ]; then
-  export MAIL="$USER@student.42.fr"
-fi
-if [ -z "$GROUP" ]; then
-  export GROUP=$(id -gn $USER)
+export USER="$(id -un)"
+export GROUP=$(id -gn $USER)
+if [ ! $AT_42 ]; then
+    export FULLNAME="fullname_placeholder"
+    export MAIL="mail_placeholder"
+else
+    export FULLNAME="$(id -F)"
+    export MAIL="$USER@student.42.fr"
 fi
 
 ######################
 #        PATHS       #
 ######################
 
-export HOMEBREW_CACHE="${HOME}/.tmp/brew_cache"
+if [ $AT_42 ]; then
+    export HOMEBREW_CACHE="${HOME}/.tmp/brew_cache"
+fi
 
 # NodeJs
-export NPM_PACKAGES=${HOME}/.npm-packages
-export NODE_PATH="${NPM_PACKAGES}/lib/node_modules:${NODE_PATH}"
+if [ "${USER}" != "root" ]; then
+    export NPM_PACKAGES=${HOME}/.npm-packages
+    export NODE_PATH="${NPM_PACKAGES}/lib/node_modules:${NODE_PATH}"
+fi
 
-export PATH="${HOME}/.brew/bin:${NPM_PACKAGES}/bin:${HOME}/.meteor:/usr/local/munki:/opt/X11/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin"
+if [ $AT_42 ]; then
+    PATH="${HOME}/.brew/bin:${NPM_PACKAGES}/bin:${HOME}/.meteor:/usr/local/munki:/opt/X11/bin"
+else
+    PATH="${HOME}/bin:${NPM_PACKAGES}/bin:${HOME}/.meteor"
+fi
+export PATH="$PATH:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin"
 #export MANPATH="${MANPATH}"
 
 if  [ -n "$(whence nvim 2>/dev/null)" ] ; then
-  export EDITOR="nvim"
+    export VI='nvim'
 else
-  export EDITOR="vim"
+    export VI='vim'
 fi
+
+export EDITOR="${VI}"
 
 ######################
 #      CONF ZSH      #
@@ -60,10 +72,10 @@ autoload -U compinit && compinit
 zstyle ':completion:*' menu select
 # add completion provied by bin installed via brew
 if [[ -d "${HOME}/.brew/share/zsh/site-functions" ]]; then
-  fpath=(${HOME}/.brew/share/zsh/site-functions $fpath)
+    fpath=(${HOME}/.brew/share/zsh/site-functions $fpath)
 fi
 if [ -f /usr/lib/node_modules/npm/lib/utils/completion.sh ]; then
-  source /usr/lib/node_modules/npm/lib/utils/completion.sh
+    source /usr/lib/node_modules/npm/lib/utils/completion.sh
 fi
 
 # make cd implicit to move into directories
@@ -102,27 +114,24 @@ autoload -U colors && colors
 ######################
 
 case "$TERM" in
-  xterm*|rxvt*|screen|vt100)
-    COLOR_TERM=$TERM;; # we know these terms have proper color support
-  *)
-    if [ -x /usr/bin/tput ] && tput setaf 1 >&/dev/null; then
-      COLOR_TERM=$TERM # we seem to have color support:
-      # assume it's compliant with ECMA-48 (ISO/IEC-6429)
-    fi;;
-esac # >>>
+    xterm*|rxvt*|screen|vt100)
+        COLOR_TERM=$TERM;; # we know these terms have proper color support
+    *)
+        if [ -x /usr/bin/tput ] && tput setaf 1 >&/dev/null; then
+            COLOR_TERM=$TERM # we seem to have color support:
+            # assume it's compliant with ECMA-48 (ISO/IEC-6429)
+        fi;;
+esac
 
-# enable color support for ls and grep when possible <<<
+# enable color support for ls and grep when possible
 if [ -n "$COLOR_TERM" ]; then
-  # export GREP_COLOR='1;32'
-  alias grep='grep --color=auto'
-  # find the option for using colors in ls: GNU or BSD version
-  unalias ls 2>/dev/null
-  ls --color -d . &>/dev/null 2>&1 \
-    && alias ls='ls --group-directories-first --color=auto' \
-    || alias ls='ls -G' # BSD
-else
-  alias ls='ls --group-directories-first'
-fi # >>>
+    alias grep='grep --color=auto'
+    # find the option for using colors in ls: GNU or BSD version
+    unalias ls 2>/dev/null
+    ls --color -d . &>/dev/null 2>&1 \
+        && alias ls='ls --group-directories-first --color=auto' \
+        || alias ls='ls -G' # BSD
+fi
 
 # ls
 unalias l 2>/dev/null
@@ -134,28 +143,35 @@ alias l='ls -lA'
 alias ll='ls -l'
 alias la='ls -A'
 
-if  [ -n "$(whence nvim 2>/dev/null)" ] ; then
-  alias vi='nvim'
-  alias vim='nvim'
-  alias neovim='nvim'
-  alias vimdiff='nvim -d'
-  alias view='nvim -R'
-fi
+alias tree='tree -a -I ".git|node_modules"'
+
+alias vi="${VI}"
+alias vim="${VI}"
+alias vimdiff="${VI} -d"
+alias view="${VI} -R"
+
+# VirtualenvWrapper
+alias mkvenv2='mkvirtualenv -p $(type -p python2)'
+alias mkvenv3='mkvirtualenv -p $(type -p python3)'
 
 # gcc
 #alias gdb='lldb'
-alias gg="gcc -Wall -Wextra -Werror"
-alias g+="g++ -Wall -Wextra -Werror"
+alias gg='gcc -Wall -Wextra -Werror'
+alias g+='g++ -Wall -Wextra -Werror'
 
 # git
-alias ga="git add"
-alias gb="git branch"
-alias gcm="git commit -m"
-alias gco="git checkout"
-alias gpl="git pull"
-alias gps="git push"
-alias gm="git merge"
-alias gu="git add -u"
+alias ga='git add'
+alias gu='git add -u'
+alias gcm='git commit -m'
+alias gbs='git status'
+alias gb='git branch'
+alias gl='git log --oneline'
+alias gcb='git checkout -b'
+alias gco='git checkout'
+alias gch='git checkout'
+alias gpl='git pull'
+alias gps='git push'
+alias gm='git merge'
 
 # smart SSH agent: http://beyond-syntax.com/blog/2012/01/on-demand-ssh-add/
 #       (see also: https://gist.github.com/1998129)
@@ -164,15 +180,16 @@ alias gu="git add -u"
 #alias git-pull='( ssh-add -l > /dev/null || ssh-add ) && git pull'
 #alias git-fetch='( ssh-add -l > /dev/null || ssh-add ) && git fetch'
 
-alias wether='curl -s http://wttr.in/Paris | head -17'
+alias weather='curl -s http://wttr.in/Paris | head -17'
 
 # tmux with 256 colors
 alias tmux="tmux -2"
 
 # cleanup
-alias vim_clean='find "${HOME}/.tmp/nvim" -mindepth 1 -delete'
-alias npm_clean='find "${HOME}/.npm-packages" -mindepth 1 -delete'
-alias redshift_stop='kill $(pgrep redshift)'
+alias vim_clean='find "${HOME}/.tmp/${VI}" -mindepth 1 -delete'
+if [ "${USER}" != "root" ]; then
+    alias npm_clean='find "${HOME}/.npm-packages" -mindepth 1 -delete'
+fi
 
 ######################
 #    PROMPT ZSH      #
@@ -185,7 +202,7 @@ zstyle ':vcs_info:*'  formats $PS_vcsinfo" [‡ %b]"
 zstyle ':vcs_info:hg*'  formats $PS_vcsinfo" [☿ %b]"
 zstyle ':vcs_info:git*' formats $PS_vcsinfo" [± %b]"
 precmd() {
-  vcs_info
+    vcs_info
 }
 
 # prompt parts
@@ -196,21 +213,20 @@ local PS_user="%n$PS_host"
 local PS_cwd="%~"
 local PS_vcsinfo="%r/%S"
 
-# comment this line for a distraction-free prompt <<<
+# comment this line for a distraction-free prompt
 local color_prompt=$COLOR_TERM
 if [ "$color_prompt" ]; then
-  autoload colors && colors
-  PS_prompt="%{$fg_bold[white]%}$PS_prompt%{$reset_color%}"
-  PS_host="%{$fg_bold[green]%}$PS_host%{$reset_color%}"
-  # time: color coded by last return code
-  PS_time="%(?.%{$fg[green]%}.%{$fg[red]%})$PS_time%{$reset_color%}"
-  # user: color coded by privileges
-  PS_user="%(!.%{$fg_bold[red]%}.%{$fg_bold[green]%})$PS_user%{$reset_color%}"
-  PS_cwd="%{$fg_bold[blue]%}$PS_cwd%{$reset_color%}"
-  PS_vcsinfo="%{$fg[blue]%}$PS_vcsinfo%{$reset_color%}"
+    autoload colors && colors
+    PS_prompt="%{$fg_bold[white]%}$PS_prompt%{$reset_color%}"
+    PS_host="%{$fg_bold[green]%}$PS_host%{$reset_color%}"
+    # time: color coded by last return code
+    PS_time="%(?.%{$fg[green]%}.%{$fg[red]%})$PS_time%{$reset_color%}"
+    # user: color coded by privileges
+    PS_user="%(!.%{$fg_bold[red]%}.%{$fg_bold[green]%})$PS_user%{$reset_color%}"
+    PS_cwd="%{$fg_bold[blue]%}$PS_cwd%{$reset_color%}"
+    PS_vcsinfo="%{$fg[blue]%}$PS_vcsinfo%{$reset_color%}"
 fi
 unset color_prompt
-# >>>
 
 # two-line prompt with time + current vcs branch on the right
 setopt prompt_subst
@@ -259,15 +275,15 @@ export LSCOLORS="$DIR$SYMLNK$SOCKET$PIPE$EXEC$BLCKSPE$CHARSPE$EXECSUID$EXECSGID$
 
 man()
 {
-  env \
-    LESS_TERMCAP_mb=$(printf "\e[1;31m") \
-    LESS_TERMCAP_md=$(printf "\e[1;31m") \
-    LESS_TERMCAP_me=$(printf "\e[0m") \
-    LESS_TERMCAP_se=$(printf "\e[0m") \
-    LESS_TERMCAP_so=$(printf "\e[1;44;33m") \
-    LESS_TERMCAP_ue=$(printf "\e[0m") \
-    LESS_TERMCAP_us=$(printf "\e[1;32m") \
-    man "$@"
+    env \
+        LESS_TERMCAP_mb=$(printf "\e[1;31m") \
+        LESS_TERMCAP_md=$(printf "\e[1;31m") \
+        LESS_TERMCAP_me=$(printf "\e[0m") \
+        LESS_TERMCAP_se=$(printf "\e[0m") \
+        LESS_TERMCAP_so=$(printf "\e[1;44;33m") \
+        LESS_TERMCAP_ue=$(printf "\e[0m") \
+        LESS_TERMCAP_us=$(printf "\e[1;32m") \
+        man "$@"
 }
 
 ######################
@@ -280,51 +296,54 @@ man()
 # VirtualenvWrapper
 export WORKON_HOME="${HOME}/.envs"
 if [ -f "${HOME}/.brew/bin/virtualenvwrapper.sh" ]; then
-  source  "${HOME}/.brew/bin/virtualenvwrapper.sh"
+    source  "${HOME}/.brew/bin/virtualenvwrapper.sh"
 fi
 
 #######################
 #   START PROGRAMS    #
 #######################
 
-if [ -n "$(whence redshift)" -a -z "$(pgrep redshift)" ] ; then
-  echo "launching redshift"
-  nohup redshift 2>&1 >/dev/null &
+if [ -n "$(whence redshift >/dev/null)" -a -z "$(pgrep redshift)" ] ; then
+    echo "launching redshift"
+    nohup redshift 2>&1 >/dev/null &
+    alias redshift_stop='kill $(pgrep redshift)'
 fi
 
-#######################
-#   EXERCICES AT 42   #
-#######################
+if [ $AT_42 ]; then
+    #######################
+    #   EXERCICES AT 42   #
+    #######################
 
-function next ()
-{
-  nb=$(basename `pwd` | grep "ex")
-  if [[ -n "$nb" ]]; then
-    if [[ -n "$1" ]]; then inc=$1; else inc=1; fi
-    nb=$(expr `echo $nb | tr -d "[a-z]"` + $inc)
-    if [[ $nb -lt 10 ]] ; then
-      dir="../ex0$nb"
-    else
-      dir="../ex$nb"
-    fi
-    mkdir -p $dir
-    cd $dir
-  fi
-}
+    function next ()
+    {
+        nb=$(basename `pwd` | grep "ex")
+        if [[ -n "$nb" ]]; then
+            if [[ -n "$1" ]]; then inc=$1; else inc=1; fi
+            nb=$(expr `echo $nb | tr -d "[a-z]"` + $inc)
+            if [[ $nb -lt 10 ]] ; then
+                dir="../ex0$nb"
+            else
+                dir="../ex$nb"
+            fi
+            mkdir -p $dir
+            cd $dir
+        fi
+    }
 
-function prev ()
-{
-  nb=$(basename `pwd` | grep "ex")
-  if [[ -n "$nb" ]]; then
-    if [[ -n "$1" ]]; then dec=$1; else dec=1; fi
-    nb=$(expr `echo $nb | tr -d "[a-z]"` - $dec)
-    if [[ $nb -lt 0 ]] ; then
-      dir="../ex00"
-    elif [[ $nb -lt 10 ]] ; then
-      dir="../ex0$nb"
-    else
-      dir="../ex$nb"
-    fi
-    cd $dir
-  fi
-}
+    function prev ()
+    {
+        nb=$(basename `pwd` | grep "ex")
+        if [[ -n "$nb" ]]; then
+            if [[ -n "$1" ]]; then dec=$1; else dec=1; fi
+            nb=$(expr `echo $nb | tr -d "[a-z]"` - $dec)
+            if [[ $nb -lt 0 ]] ; then
+                dir="../ex00"
+            elif [[ $nb -lt 10 ]] ; then
+                dir="../ex0$nb"
+            else
+                dir="../ex$nb"
+            fi
+            cd $dir
+        fi
+    }
+fi
